Added MS Paint
This takes a while so instead of using OpenCV to make a video and record that, we use a frame sequence and then take a screenshot of each frame, and render that to an image. We can then import this into an editing software (in my case Blender) and add the video.